Update CommonCLI.cpp
I keep wanting a help menu in the CLI but it isn't there.  Had a look and realised a small refactor would help make this maintainable.  I have:

- made a single source of truth for the output strings
- added a number of formatter functions to enable small help responses (essentially grouped things)
- added the calling routines for the above into the main logic

CLI message responses for use over the mesh need to be <= 160 chars which I *think* I have achieved here.  Needed to group things so that it didn't just try to dump everything in one hit and go way over that 160 limit.  Was also conscious of future expansion so I hope these messages are still small enough for additional options in the future.

I believe this is a direct drop-in with no other files / code affected.  This is first commit after development though and has not yet been tested on a device.